Moved state and imports into Calculator component
